Settings: Update AppNotificationSettings if user resumes
Right now if a user resumes this activty (eg from recents) and did change in the
mean time settings which effects the visibility of preferences in AppNotificationSettings
(eg were removed) it does not get updated.

Fix it with simply reloading the content on onResume and decide to show the entries
on the current user settings situation.

Change-Id: I0b16ebf35604fdca1611f026c9b609427f3060c8

Conflicts:
	src/com/android/settings/notification/AppNotificationSettings.java